refactor: remove unreachable runtime mismatch defer queue
