+++ src/services/discord/router/mod.rs
@@ -26,6 +26,7 @@ pub(super) use message_handler::{
 };
 pub(crate) use message_handler::{IntakeRequest, execute_intake_turn_core};
 pub(in crate::services::discord) use queue_status_presentation::queue_status_card_enabled;
+pub(super) use turn_start::reserve_headless_turn;
 pub(crate) use turn_start::{
     HeadlessTurnReservation, HeadlessTurnStartError, HeadlessTurnStartOutcome,
     HeadlessTurnStartStatus,
@@ -34,7 +35,6 @@ pub(crate) use turn_start::{
 pub(crate) use turn_start::{
     load_session_runtime_state, try_intake_runtime_transition_after_redirect,
 };
-pub(super) use turn_start::{reserve_headless_turn, reserve_headless_turn_with_user_msg_id};
 
 // Re-export items used across submodules
 use thread_binding::{link_dispatch_thread, lookup_dispatch_info, verify_thread_accessible};
